+++ test_features.py
@@ -1,0 +1,17 @@
+from features import *
+
+sentences = [
+    "Hello World" ,
+    "Hello World from Java" ,
+    "Ashwin is a hardworking boy of age 20." ,
+    "Hello World from Python" ,
+    "Hello World" ,
+    "Hello World from Java" ,
+    "Ashwin is a hardworking boy of age 20."
+    "Hello World from Python"
+]
+
+print( sentence_length( sentences ) )
+print( sentence_position( sentences ) )
+print( sentence_num_numeric_terms( sentences ) )
+print( sentence_num_proper_nouns( sentences ) )