+++ test_keyword_extraction.py
@@ -1,0 +1,12 @@
+import yake
+
+kw_extractor = yake.KeywordExtractor()
+text = """spaCy is an open-source software library for advanced natural language processing, written in the programming languages Python and Cython. The library is published under the MIT license and its main developers are Matthew Honnibal and Ines Montani, the founders of the software company Explosion."""
+language = "en"
+max_ngram_size = 3
+deduplication_threshold = 0.9
+numOfKeywords = 20
+custom_kw_extractor = yake.KeywordExtractor(lan=language, n=max_ngram_size, dedupLim=deduplication_threshold, top=numOfKeywords, features=None)
+keywords = custom_kw_extractor.extract_keywords(text)
+for kw in keywords:
+    print(kw)