+++ CHANGELOG.md
@@ -1,2 +1,4 @@
-- Updated `sequential.py` for `Trial 2`
-- Added `trail_1_summaries.zip`
+- Created `trial_03.py` for testing Jaccard coefficient as the similarity metric
+- Added `test_keyword_extraction.py` to test keyword extraction feature
+- Added `compute_scores.py` to determine ROUGE scores from `summaries`
+- Added results for `trial_03`